requirements.txt
openai>=1.50.0
anthropic>=0.36.0
httpx>=0.27.0
python-dotenv>=1.0.0
